+++ src/pdevices/nativepdevice.cpp
@@ -33,7 +33,7 @@ using namespace luxrays;
 size_t NativePixelDevice::SampleBufferSize = 4096;
 
 NativePixelDevice::NativePixelDevice(const Context *context,
-		const unsigned int threadIndex, const unsigned int devIndex) :
+		const size_t threadIndex, const size_t devIndex) :
 			PixelDevice(context, DEVICE_TYPE_NATIVE_THREAD, devIndex) {
 	char buf[64];
 	sprintf(buf, "NativePixelThread-%03d", (int)threadIndex);
